Refactor familiar structure, add config, update flute and cheese recipe
