+++ gradle.properties
@@ -13,7 +13,7 @@ loom_version=1.15-SNAPSHOT
 yarn_mappings=1.20.1+build.10
 
 # Mod Properties
-mod_version=0.1.2
+mod_version=0.1.3
 maven_group=com.rnoobb.rats
 archives_base_name=rats-and-creatures
