+++ src/client/java/com/rnoobb/rats/client/renderer/layer/RatArmorLayer.java
@@ -1,6 +1,7 @@
 package com.rnoobb.rats.client.renderer.layer;
 
 import com.rnoobb.rats.entity.custom.RatEntity;
+import com.rnoobb.rats.entity.custom.AbstractHelperEntity;
 import net.minecraft.client.render.OverlayTexture;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumer;
