+++ src/client/java/com/rnoobb/rats/client/renderer/layer/RatItemLayer.java
@@ -1,6 +1,7 @@
 package com.rnoobb.rats.client.renderer.layer;
 
 import com.rnoobb.rats.entity.custom.RatEntity;
+import com.rnoobb.rats.entity.custom.AbstractHelperEntity;
 import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.render.model.json.ModelTransformationMode;
 import net.minecraft.client.util.math.MatrixStack;
