+++ src/main/java/com/rnoobb/rats/mixin/CatEntityMixin.java
@@ -1,6 +1,7 @@
 package com.rnoobb.rats.mixin;
 
 import com.rnoobb.rats.entity.custom.RatEntity;
+import com.rnoobb.rats.entity.custom.AbstractHelperEntity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.ai.goal.ActiveTargetGoal;
 import net.minecraft.entity.ai.goal.MeleeAttackGoal;
