+++ src/main/resources/data/rats_and_creatures/recipes/cheese_craft.json
@@ -1,5 +1,5 @@
 {
-  "type": "minecraft:smoking",
+  "type": "minecraft:smelting",
   "ingredient": {
     "item": "minecraft:milk_bucket"
   },
